Split storage bucket methods
